requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.3.1
streamlit==1.38.0
python-dotenv==1.0.1